test(transpiler): make test more reliable

+++ packages/transpiler/__test__/utils/setup.ts
@@ -7,5 +7,7 @@ const buildFolder = path.join(__dirname, '../..', '.build/src')
 process.env.BEARER_SCENARIO_ID = 'SPONGE_BOB'
 
 globby.sync(path.join(buildFolder, '/**/*.tsx')).forEach(file => {
-  fs.unlinkSync(file)
+  if (fs.existsSync(file)) {
+    fs.unlinkSync(file)
+  }
 })
